# expression: implement vectorized evaluation for `builtinBenchmarkSig`