PCP #12103 

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinBenchmarkSig` at #12103 

### What is changed and how it works?
It gets faster about 60x.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
```
> GO111MODULE=on go test -check.f TestVectorizedBuiltinInfoFunc
PASS: builtin_info_vec_test.go:98: testEvaluatorSuite.TestVectorizedBuiltinInfoFunc	0.038s
OK: 1 passed
PASS
ok  	github.com/pingcap/tidb/expression	0.064s
```

 - Benchmark
```
> go test -v -benchmem -bench=BenchmarkVectorizedBuiltinInfoFunc -run=BenchmarkVectorizedBuiltinInfoFunc -args builtinBenchmarkSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinInfoFunc/builtinBenchmarkSig-VecBuiltinFunc-12         	  970549	      1209 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinInfoFunc/builtinBenchmarkSig-NonVecBuiltinFunc-12      	   15642	     74983 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	4.149s
```